backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-anthropic
redis
langchain-redis
pydantic
httpx
gunicorn
